refactor: normalize package helpers
