+++ src/helpers/filters.ts
@@ -1,5 +1,4 @@
-import type { GitHubClient } from "../clients";
-import type { error, GitHub } from "../types";
+import type { error, GitHub, GitHubClient } from "./types";
 
 import filters from "../filters";
 import { EVENT_PULL_REQUEST, EVENT_PUSH } from "./constants";
@@ -9,6 +8,8 @@ interface Config {
     readonly value: string[];
   };
 }
+
+type Filter = (s: string) => boolean;
 
 interface Params {
   readonly client: GitHubClient;
@@ -20,8 +21,8 @@ async function getFilters({
   client,
   config,
   github,
-}: Params): Promise<[((s: string) => boolean)[], error]> {
-  const { context } = github;
+}: Params): Promise<[Filter[], error]> {
+  const context = github.context;
   const event = context.eventName;
 
   const result = [
