+++ src/helpers/types.d.ts
@@ -1,0 +1,8 @@
+import type { GitHubClient } from "../clients";
+import type { error, GitHub } from "../types";
+
+export type {
+  error,
+  GitHub,
+  GitHubClient,
+};
